@@ -27,7 +27,7 @@ classifiers =
 package_dir =
     = src
 packages = find:
-python_requires = >=3.4
+python_requires = >=3.6
 install_requires =
     Babel>=2.3.4
     iso8601>=0.1.11
@@ -39,10 +39,13 @@ where = src
 
 [options.extras_require]
 dev =
+    black
     coverage
     docformatter
     flake8
+    flake8-black
     flake8-bugbear
+    flake8-isort
     invoke
     isort
     pylint
