@@ -15,9 +15,9 @@ from invoke import Exit, UnexpectedExit, run as _run, task
 
 
 PACKAGE_NAME = "zulu"
-PACKAGE_SOURCE = "src/{}".format(PACKAGE_NAME)
-TEST_TARGETS = "{} tests".format(PACKAGE_SOURCE)
-LINT_TARGETS = "{} tasks.py".format(TEST_TARGETS)
+PACKAGE_SOURCE = f"src/{PACKAGE_NAME}"
+TEST_TARGETS = f"{PACKAGE_SOURCE} tests"
+LINT_TARGETS = f"{TEST_TARGETS} tasks.py"
 EXIT_EXCEPTIONS = (Exit, UnexpectedExit, SystemExit)
 
 
@@ -28,23 +28,21 @@ run = partial(_run, pty=True)
 @task
 def black(ctx, quiet=False):
     """Autoformat code using black."""
-    run("black {}".format(LINT_TARGETS), hide=quiet)
+    run(f"black {LINT_TARGETS}", hide=quiet)
 
 
 @task
 def isort(ctx, quiet=False):
     """Autoformat Python imports."""
-    run("isort {}".format(LINT_TARGETS), hide=quiet)
+    run(f"isort {LINT_TARGETS}", hide=quiet)
 
 
 @task
 def docformatter(ctx):
     """Autoformat docstrings using docformatter."""
     run(
-        "docformatter -r {} "
-        "--in-place --pre-summary-newline --wrap-descriptions 100 --wrap-summaries 100".format(
-            LINT_TARGETS
-        )
+        f"docformatter -r {LINT_TARGETS} "
+        f"--in-place --pre-summary-newline --wrap-descriptions 100 --wrap-summaries 100"
     )
 
 
@@ -64,13 +62,13 @@ def fmt(ctx):
 @task
 def flake8(ctx):
     """Check code for PEP8 violations using flake8."""
-    run("flake8 --format=pylint {}".format(LINT_TARGETS))
+    run(f"flake8 --format=pylint {LINT_TARGETS}")
 
 
 @task
 def pylint(ctx):
     """Check code for static errors using pylint."""
-    run("pylint {}".format(LINT_TARGETS))
+    run(f"pylint {LINT_TARGETS}")
 
 
 @task
@@ -80,7 +78,7 @@ def lint(ctx):
     failures = []
 
     for name, linter in linters.items():
-        print("Running {}".format(name))
+        print(f"Running {name}")
         try:
             linter(ctx)
         except EXIT_EXCEPTIONS:
@@ -88,15 +86,15 @@ def lint(ctx):
             result = "FAILED"
         else:
             result = "PASSED"
-        print("{}\n".format(result))
+        print(f"{result}\n")
 
     if failures:
         failed = ", ".join(failures)
-        raise Exit("ERROR: Linters that failed: {}".format(failed))
+        raise Exit(f"ERROR: Linters that failed: {failed}")
 
 
 @task(help={"args": "Override default pytest arguments"})
-def unit(ctx, args="{} --cov={} --flake8 --pylint".format(TEST_TARGETS, PACKAGE_NAME)):  # noqa
+def unit(ctx, args=f"{TEST_TARGETS} --cov={PACKAGE_NAME} --flake8 --pylint"):
     """Run unit tests using pytest."""
     tox_env_site_packages_dir = os.getenv("TOX_ENV_SITE_PACKAGES_DIR")
     if tox_env_site_packages_dir:
@@ -104,7 +102,7 @@ def unit(ctx, args="{} --cov={} --flake8 --pylint".format(TEST_TARGETS, PACKAGE_
         tox_env_pkg_src = os.path.join(tox_env_site_packages_dir, os.path.basename(PACKAGE_SOURCE))
         args = args.replace(PACKAGE_SOURCE, tox_env_pkg_src)
 
-    run("pytest {}".format(args))
+    run(f"pytest {args}")
 
 
 @task
@@ -120,7 +118,7 @@ def test(ctx):
     lint(ctx)
 
     print("Running unit tests")
-    unit(ctx, args="{} --cov={}".format(TEST_TARGETS, PACKAGE_NAME))
+    unit(ctx, args=f"{TEST_TARGETS} --cov={PACKAGE_NAME}")
 
 
 @task
@@ -130,11 +128,8 @@ def docs(ctx, serve=False, bind="127.0.0.1", port=8000):
     run("sphinx-build -q -W -b html docs docs/_build/html")
 
     if serve:
-        print("Serving docs on {0} port {1} (http://{0}:{1}/) ...".format(bind, port))
-        run(
-            "python -m http.server -b {} --directory docs/_build/html {}".format(bind, port),
-            hide=True
-        )
+        print(f"Serving docs on {bind} port {port} (http://{bind}:{port}/) ...")
+        run(f"python -m http.server -b {bind} --directory docs/_build/html {port}", hide=True)
 
 
 @task
